Update CS on the tests. Move test stubs to own file.
